fix: FormApplicationListQuery 클래스를 MyFormApplicationPageQuery의 이너 클래스로 수정
