+++ src/main/java/ddingdong/ddingdongBE/domain/formapplication/controller/dto/response/MyFormApplicationPageResponse.java
@@ -1,7 +1,7 @@
 package ddingdong.ddingdongBE.domain.formapplication.controller.dto.response;
 
-import ddingdong.ddingdongBE.domain.formapplication.service.dto.query.FormApplicationListQuery;
 import ddingdong.ddingdongBE.domain.formapplication.service.dto.query.MyFormApplicationPageQuery;
+import ddingdong.ddingdongBE.domain.formapplication.service.dto.query.MyFormApplicationPageQuery.FormApplicationListQuery;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
